requirements.txt
Flask>=2.0.0
flask-cors>=4.0.0
groq>=0.9.0,<0.11.0
gTTS>=2.5.0
gunicorn>=21.2.0
python-dotenv>=1.0.0